Teleport killed ghosts in the middle of the ghost room
The position was hard-coded because in the teacher's specifications it
says that they must spawn in the room and not leave as long as pacman is
invincible. Because of this we can't respawn the ghosts at their starting
position because the red ghost is initially outside of the ghost room.